requirements.txt
langchain==0.3.7
langchain-openai==0.2.9
langchain-community==0.3.7
openai==1.55.3
python-dotenv==1.0.1
streamlit==1.41.1
faiss-cpu==1.8.0
pypdf==4.3.1
tiktoken==0.7.0